Major Refactor - Navbar Component
