+++ frontend/src/app.js
@@ -15,13 +15,17 @@ import AddBroadcasts from "./components/admin/add_broadcasts/add_broadcasts";
 import Faqs from "./components/faq/faq";
 import NotFound from "./components/404/not_found";
 import Login from "./components/auth/login/login";
-import Navbar from "./components/navbar/navbar";
 import Footer from "./components/footer/footer";
 import Terms from "./components/terms/terms";
 import GetInvolved from "./components/get_involved/get_involved";
 import PrivacyPolicy from "./components/privacy-policy/privacy-policy";
 import Setting from "./components/admin/setting/setting";
 import "./app.css";
+
+// Component
+import { Navbar } from "./components/Navbar/index";
+
+// Pages
 
 const App = () => {
   return (
