+++ frontend/src/components/Navbar/navbar.module.scss
